+++ tests/testthat/test11_vmss.R
@@ -11,6 +11,9 @@ if(tenant == "" || app == "" || password == "" || subscription == "")
 vmss_name <- paste0("vmss", paste0(sample(letters, 10, TRUE), collapse=""))
 location <- "australiaeast"
 
+# turn off parallelisation
+maxpoolsize <- options(azure_vm_maxpoolsize=0)
+
 rg <- AzureRMR::az_rm$
     new(tenant=tenant, app=app, password=password)$
     get_subscription(subscription)$
@@ -18,7 +21,7 @@ rg <- AzureRMR::az_rm$
 
 test_that("Scaleset creation works",
 {
-    vm <- rg$create_vm_scaleset(vmss_name, user_config("username", "../resources/testkey.pub"), instances=2)
+    vm <- rg$create_vm_scaleset(vmss_name, user_config("username", "../resources/testkey.pub"), instances=3)
     expect_is(vm, "az_vmss_template")
 })
 
@@ -27,11 +30,18 @@ test_that("Scaleset interaction works",
     vm <- rg$get_vm_scaleset(vmss_name)
     expect_is(vm, "az_vmss_template")
 
+    expect_is(vm$get_public_ip_address(), "character")
+
+    inst <- vm$list_instances()
+    expect_is(inst, "list")
+
     expect_silent(vm$run_script("ls /tmp"))
 
     expect_is(vm$get_vm_private_ip_addresses(), "character")
     expect_is(vm$get_vm_public_ip_addresses(), "character")
-    expect_is(vm$get_public_ip_address(), "character")
+
+    expect_is(vm$get_vm_private_ip_addresses(names(inst)[1:2]), "character")
+    expect_is(vm$get_vm_public_ip_addresses(names(inst)[1:2]), "character")
 })
 
 test_that("Scaleset deletion works",
@@ -44,3 +54,4 @@ test_that("Scaleset deletion works",
 })
 
 rg$delete(confirm=FALSE)
+options(maxpoolsize)